refactor(checker): split indexed access type nodes
